+++ src/tables/mod.rs
@@ -11,6 +11,7 @@ pub mod cmap;
 pub mod cmap_uvs;
 pub mod colr;
 pub mod context;
+pub mod cpal;
 pub mod device;
 pub mod fvar;
 pub mod gdef;